Parallelize build-time meta-image rendering with worker threads
Extract the card templates and Satori/resvg rendering from
generate-meta-images.mjs into scripts/meta-images/render.mjs so the
generator and a new worker_threads entry (render-worker.mjs) share one
definition. Cold runs (fresh checkout or OG_TEMPLATE_VERSION bump) fan
out across a worker pool sized to the machine; small incremental
batches (< 8 cards) or OG_WORKERS=1 render inline to skip worker
startup. No visual change — the moved template code is byte-identical.
